requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
roboflow>=1.1.0
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
Pillow>=8.0.0
pyyaml>=6.0
